backend/requirements.txt
Django==3.1.3
djoser==2.1.0
psycopg2==2.8.6
django-filter==2.4.0
django-rest-swagger==2.2.0
djangorestframework==3.12.2
